requirements.txt
python-pptx==0.6.21
pyinstaller==6.3.0
